requirements.txt
face_recognition==1.3.0
gradio==4.44.1
Pillow==10.4.0
Requests==2.32.3
sentence_transformers==3.1.1
